backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
langchain
langchain-community
langchain-groq
langchain-core
langchain-chroma
langchain-huggingface
langchain-text-splitters
chromadb
fastembed
pymupdf
python-multipart
google-generativeai
langgraph
tavily-python